Move Query to its own file
